requirements.txt
numpy
visdom
Pillow
tqdm
torch
tensorflow-tensorboard
tensorboard-pytorch
